requirements.txt
langchain-google-genai~=2.0.11
langchain-openai~=0.3.1
pydantic~=2.10.6
pytest~=8.3.4
asyncio~=3.4.3
playwright~=1.50.0
langchain-core~=0.3.40
python-dotenv~=1.0.1